+++ src/main/java/net/servicestack/idea/php/PhpNativeTypesHandler.java
@@ -1,0 +1,36 @@
+package net.servicestack.idea.php;
+
+import net.servicestack.idea.common.BaseNativeTypesHandler;
+import net.servicestack.idea.common.NativeTypesLanguage;
+
+public class PhpNativeTypesHandler extends BaseNativeTypesHandler {
+    @Override
+    public String getFileExtension() {
+        return ".php";
+    }
+
+    @Override
+    public String getRelativeTypesUrl() {
+        return "types/php";
+    }
+
+    @Override
+    public NativeTypesLanguage getTypesLanguage() {
+        return NativeTypesLanguage.Php;
+    }
+
+    @Override
+    public String getOptionsCommentStart() {
+        return "<?php namespace dtos;\n/* Options:";
+    }
+
+    @Override
+    public String getOptionsCommentEnd() {
+        return "*/";
+    }
+
+    @Override
+    public String getOptionsIgnoreFlag() {
+        return "//";
+    }
+}